Bug 466260 - OutOfMemoryError in Parser.createStringLiteral (4582)
JSDT Core part of the fix:
The files that are the children of a 'derived' folder are excluded from the build and threated as read-only.
The 'derived' flag is usually set by the maven import in order to indicate that the resource (some target folder contents) can
be removed upon a build operation, so it shouldn't be treated as a source.

Signed-off-by: vrubezhny <vrubezhny@exadel.com>